Replace connection type in `App` function signature
There's a bit more work involved when running tests in parallel against
a sqlite DB. The test suite needs to keep track of open connections
operating against the DB as only one operation is allowed at a time. If
the DB is busy, attempts to read or write will panic on `Database is
locked` error; thus, connections need to be stateful. We can wait for
the next available connection using a shared connection pool across
tests that need to open a DB connection.

This neccesitates the following changes:

* Update the function signatures to the appropriate connection type;
* Add a stateful connection pool in the tests via `OnceLock` API;
* Change the main function to use a connection from a pool, based
  on the signature updates.